requirements/base.txt
# Django
Django
django-authtools
djangorestframework
django-admin-list-filter-dropdown

# Database
psycopg2
dj-database-url

# Documentation
Sphinx
sphinx-autobuild
sphinx-rtd-theme

# Miscellaneous
xlwt
fpdf
